refactored dir structure
