--- src/app/Resume/ResumeContainer.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import Resume from './components/Resume';
-
-// TODO: Add tests for component
-
-function ResumeContainer() {
-  return (
-    <Resume />
-  );
-}
-
-ResumeContainer.title = 'resume';
-
-export default ResumeContainer;
