--- src/app/Resume/stylesheets/Resume.scss
@@ -1,21 +1,0 @@
-.resume {
-  font-size: 14px;
-  color: #777;
-}
-
-.download-option {
-  display: inline-block;
-  padding-right: 10px;
-}
-
-.resume-section {
-  margin-bottom: 20px;
-}
-
-.resume-education {
-  display: block;
-}
-
-.resume-list {
-  padding-left: 20px;
-}
