+++ src/app/components/ConnectedIntlProvider/ConnectedIntlProvider.js
