+++ src/app/components/MyNavBar/MyNavBar.js
@@ -37,7 +37,7 @@ const componentMessages = defineMessages({
 });
 
 const items = [
-  // { path: '/projects', label: 'projects' },
+  { path: '/projects', label: 'projects' },
 ];
 
 const navOptions = items.map(item =>
