+++ src/app/redux/configureStore.js
@@ -1,0 +1,3 @@
+const configureStore = (process.env.NODE_ENV === 'production') ? require('./store/configureStore.prod') : require('./store/configureStore.dev');
+
+module.exports = configureStore;
