+++ src/app/redux/modules/CustomNavBar/__tests__/reducers-test.js
@@ -1,7 +1,8 @@
+// TODO: Convert to jest
 import { expect } from 'chai';
 
-import updateLocale from '../actions/action_creators';
-import reducer from './reducers';
+import updateLocale from '../actionCreators';
+import reducer from '../reducers';
 
 describe('reducers/CustomNavBar', () => {
   it('handles UPDATE_LOCALE', () => {
