+++ src/app/redux/modules/CustomNavBar/constants.js
