+++ src/app/redux/modules/CustomNavBar/reducers.js
@@ -1,4 +1,4 @@
-import UPDATE_LOCALE from '../constants/actionTypes';
+import UPDATE_LOCALE from './constants';
 
 function updateLocale(state, newState) {
   return { ...state, ...newState };
