+++ src/app/redux/rootReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import auth from '../Auth/reducers/reducers';
 import todo from '../Todo/reducers/reducers';
-import intl from '../CustomNavBar/reducers/reducers';
+import intl from './modules/CustomNavBar/reducers';
 
 export default combineReducers({
   todo,
