--- src/app/store/configureStore.js
@@ -1,3 +1,0 @@
-const configureStore = (process.env.NODE_ENV === 'production') ? require('./configureStore.prod') : require('./configureStore.dev');
-
-module.exports = configureStore;
