--- src/app/stylesheets/ItemClass.scss
@@ -1,5 +1,0 @@
-.item {
-  background: #e5f0ff;
-  border-bottom: 2px solid white;
-  padding: 10 0 0 10;
-}
